pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "myproject"
version = "1.0"
dependencies = [
    "Django==5.1.*",
    "djangorestframework==3.15.*",
    "django-filter==24.3",
    "Markdown==3.7",
    "psycopg[binary]==3.2.2",
    "requests==2.32.3"
]

[project.optional-dependencies]
dev = [
  "drf-spectacular==0.27.2",
  "requests_mock==1.12.1"
]
